requirements.txt
django==3.2.7
djangorestframework==3.12.4
django-filter==2.4.0
django-cors-headers==3.8.0
djangorestframework-simplejwt==5.0.0
